pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lesspython"
version = "0.0.0"
description = "Find duplicated Python code fragments via AST hashing"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.9",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "ruff",
  "ty",
]

[project.scripts]
lesspython = "lesspython.cli:app"

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
line-length = 80
src = ["lesspython", "tests"]
exclude = ["examples", ".venv"]

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.ty]
python_version = "3.10"
strict = false

[tool.ty.src]
exclude = ["examples", ".venv"]
